job_platform/tests/test_routes.py
#!/usr/bin/python3
"""
Contains the TestRoutesDocs classes
"""

import unittest
from app import app, db
from models import User

class TestRoutes(unittest.TestCase):
    def setUp(self):
        self.app = app.test_client()
        self.app.testing = True
        app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///test.db'
        db.create_all()

        user = User(username='testuser', email='test@example.com', password='password', user_type='jobseeker')
        db.session.add(user)
        db.session.commit()

    def tearDown(self):
        db.session.remove()
        db.drop_all()

    def test_register_route(self):
        response = self.app.get('/register')
        self.assertEqual(response.status_code, 200)
        self.assertIn(b'Register', response.data)

    def test_login_route(self):
        response = self.app.get('/login')
        self.assertEqual(response.status_code, 200)
        self.assertIn(b'Login', response.data)

    def test_dashboard_route(self):
        response = self.app.get('/dashboard')
        self.assertEqual(response.status_code, 302)  # Redirect to login page if not logged in

if __name__ == '__main__':
    unittest.main()
